+++ src/main/java/devping/nnplanner/domain/openapi/batch/food/FoodItemProcessor.java
@@ -1,4 +1,4 @@
-package devping.nnplanner.domain.openapi.FoodBatch;
+package devping.nnplanner.domain.openapi.batch.food;
 
 import devping.nnplanner.domain.openapi.dto.response.FoodApiResponseDTO.FoodItem;
 import devping.nnplanner.domain.openapi.entity.Food;
